+++ CHANGES.md
@@ -65,6 +65,8 @@ Apollo 1.9.0
 * [fix size of create project button](https://github.com/ctripcorp/apollo/pull/3849)
 * [translation of "portal-how-to-enable-webhook-notification.md"](https://github.com/ctripcorp/apollo/pull/3847)
 * [feature: add history detail for not key-value type of namespace](https://github.com/ctripcorp/apollo/pull/3856)
+* [fix show-text-modal number display](https://github.com/ctripcorp/apollo/pull/3851)
+
 ------------------
 All issues and pull requests are [here](https://github.com/ctripcorp/apollo/milestone/6?closed=1)